cleanup network manager
